Refactor CrossOriginHelperInternal to internal package and rename
