+++ microprofile/bundles/helidon-microprofile/pom.xml
@@ -75,6 +75,10 @@
             <artifactId>helidon-microprofile-tracing</artifactId>
         </dependency>
         <dependency>
+            <groupId>io.helidon.microprofile</groupId>
+            <artifactId>helidon-microprofile-cors</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.glassfish.jersey.media</groupId>
             <artifactId>jersey-media-json-binding</artifactId>
             <scope>runtime</scope>
