+++ pom.xml
@@ -977,7 +977,6 @@ helidon-parent,helidon-dependencies,helidon-bom,helidon-se,helidon-mp,io.grpc,he
                                 <sourceFileExclude>**/test/**/*.java</sourceFileExclude>
                                 <sourceFileExclude>**/*_.java</sourceFileExclude>
                                 <sourceFileExclude>**/io/grpc/stub/**/*.java</sourceFileExclude>
-                                <sourceFileExclude>**/io/helidon/webserver/cors/CrossOriginHelperInternal.java</sourceFileExclude>
                             </sourceFileExcludes>
                         </configuration>
                         <reports>
