+++ webserver/cors/src/main/java/module-info.java
@@ -26,4 +26,5 @@ module io.helidon.webserver.cors {
     requires io.helidon.webserver;
 
     exports io.helidon.webserver.cors;
+    exports io.helidon.webserver.cors.internal to io.helidon.microprofile.cors;
 }
